requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
rich>=13.7.0
pydantic>=2.5.0
pyyaml>=6.0.1
streamlit>=1.36.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=23.0.0
